refactor(web): collapse search-notebook-data entry points onto /notebooks
Three surfaces previously promised "search notebook data": the
"Notebook-Daten durchsuchen" tool button on /workplace, the standalone
/research page with its own filter UI, and /notebooks + /suche. The
canonical pair is /notebooks (gallery + @notebook mentions in chat) and
/suche (SearchGraph agent); the parallel /research page duplicated the
capability and the workplace button advertised it under a misleading
label.

/recherche and /research both redirect to /notebooks now. The backend
/api/research/* router stays because NotebookManualSearch (inside the
notebook page itself) still uses it; its supporting hooks/components
moved into features/notebook/manual-search/ since they have no other
consumer now.
